+++ packages/shared/src/components/header/MobileExploreHeader.tsx
